requirements.txt
pylint
yfinance
matplotlib
python-decouple
openai
logging
argparse
numpy
praw
torch
futures
pytz
opentelemetry-api
opentelemetry-instrumentation
pyprojroot
scipy
pandas
pyzmq
pymc
arviz
arviz_base
arviz_stats
arviz_plots
seaborn
scipy
scikit-learn
tensorflow